+++ ValheimServersideQoL/CHANGELOG.md
@@ -1,3 +1,7 @@
+### v1.0.1030 (BETA)
+- Adjusted teleportation of non-teleportable items [#99](https://github.com/ArgusMagnus/ValheimServersideQoL/issues/99):
+   - Return items to players that logged out in proximity of a portal when they log back in
+
 ### v1.0.1027 (BETA)
 - Adjusted teleportation of non-teleportable items [#99](https://github.com/ArgusMagnus/ValheimServersideQoL/issues/99):
     - Made type of shown message configurable